Refactor AOV and AOVAccumulator code
- Add *protected* field where it's missing.
- Split the *diagnosticaov* file into multiple files.
- Make AOV model name variables consistent.
- Fix channel count for AOVs.
- Ignore distance to pixel's center in unfiltered AOVs.
- Handle multiple passes with Pixel Sample Count AOV.
- Handle multiple passes with Pixel Variation AOV.
- Ignore pixels outside of the crop window in the Invalid Samples AOV.
- Fill the Invalid Samples AOV in *post_process*.
- This closes #2127.
