+++ src/appleseed/renderer/modeling/aov/aov.cpp
@@ -31,6 +31,7 @@
 
 // appleseed.renderer headers.
 #include "renderer/kernel/aov/imagestack.h"
+#include "renderer/modeling/frame/frame.h"
 
 // appleseed.foundation headers.
 #include "foundation/image/color.h"
@@ -70,6 +71,15 @@ void AOV::release()
     delete this;
 }
 
+Image& AOV::get_image() const
+{
+    return *m_image;
+}
+
+void AOV::post_process_image(const Frame& frame)
+{
+}
+
 void AOV::create_image(
     const size_t        canvas_width,
     const size_t        canvas_height,
@@ -79,23 +89,10 @@ void AOV::create_image(
 {
     m_image_index = aov_images.append(
         get_name(),
-        4, // todo: check if we can pass aov->get_channel_count() here
+        get_channel_count(),
         PixelFormatFloat);
+
     m_image = &aov_images.get_image(m_image_index);
-}
-
-Image& AOV::get_image() const
-{
-    return *m_image;
-}
-
-void AOV::clear_image()
-{
-    m_image->clear(Color4f(0.0f));
-}
-
-void AOV::post_process_image(const AABB2u& crop_window)
-{
 }
 
 
@@ -110,18 +107,23 @@ ColorAOV::ColorAOV(const char* name, const ParamArray& params)
 
 size_t ColorAOV::get_channel_count() const
 {
-    return 3;
+    return 4;
 }
 
 const char** ColorAOV::get_channel_names() const
 {
-    static const char* ChannelNames[] = {"R", "G", "B"};
+    static const char* ChannelNames[] = {"R", "G", "B", "A"};
     return ChannelNames;
 }
 
 bool ColorAOV::has_color_data() const
 {
     return true;
+}
+
+void ColorAOV::clear_image()
+{
+    m_image->clear(Color4f(0.0f, 0.0f, 0.0f, 0.0f));
 }
 
 
@@ -131,19 +133,33 @@ bool ColorAOV::has_color_data() const
 
 UnfilteredAOV::UnfilteredAOV(const char* name, const ParamArray& params)
   : AOV(name, params)
-  , m_filter_image(nullptr)
 {
 }
 
 UnfilteredAOV::~UnfilteredAOV()
 {
     delete m_image;
-    delete m_filter_image;
+}
+
+size_t UnfilteredAOV::get_channel_count() const
+{
+    return 3;
+}
+
+const char** UnfilteredAOV::get_channel_names() const
+{
+    static const char* ChannelNames[] = {"R", "G", "B"};
+    return ChannelNames;
 }
 
 bool UnfilteredAOV::has_color_data() const
 {
     return false;
+}
+
+void UnfilteredAOV::clear_image()
+{
+    m_image->clear(Color3f(0.0f, 0.0f, 0.0f));
 }
 
 void UnfilteredAOV::create_image(
@@ -162,19 +178,8 @@ void UnfilteredAOV::create_image(
             get_channel_count(),
             PixelFormatFloat);
 
-    // Extra image to keep track of the distance
-    // to the nearest sample for each pixel.
-    m_filter_image =
-        new Image(
-            canvas_width,
-            canvas_height,
-            tile_width,
-            tile_height,
-            1,
-            PixelFormatFloat);
-
     // We need to clear the image because the default channel value
-    // might not be zero and also to initialize the pixel distance channel.
+    // might not be zero.
     clear_image();
 }
