+++ src/appleseed/renderer/modeling/aov/aov.h
@@ -88,10 +88,10 @@ class APPLESEED_DLLSYMBOL AOV
     foundation::Image& get_image() const;
 
     // Clear the AOV image to default values.
-    virtual void clear_image();
+    virtual void clear_image() = 0;
 
     // Apply any post processing needed to the AOV image.
-    virtual void post_process_image(const foundation::AABB2u& crop_window);
+    virtual void post_process_image(const Frame& frame);
 
   protected:
     friend class AOVAccumulatorContainer;
@@ -125,13 +125,16 @@ class ColorAOV
     ColorAOV(const char* name, const ParamArray& params);
 
     // Return the number of channels of this AOV.
-    size_t get_channel_count() const override;
+    virtual size_t get_channel_count() const override;
 
     // Return the AOV channel names.
-    const char** get_channel_names() const override;
+    virtual const char** get_channel_names() const override;
 
     // Return true if this AOV contains color data.
-    bool has_color_data() const override;
+    virtual bool has_color_data() const override;
+
+    // Clear the AOV image to default values.
+    virtual void clear_image() override;
 };
 
 
@@ -149,13 +152,20 @@ class UnfilteredAOV
     // Destructor.
     ~UnfilteredAOV() override;
 
+    // Return the number of channels of this AOV.
+    virtual size_t get_channel_count() const override;
+
+    // Return the AOV channel names.
+    virtual const char** get_channel_names() const override;
+
     // Return true if this AOV contains color data.
-    bool has_color_data() const override;
+    virtual bool has_color_data() const override;
+
+    // Clear the AOV image to default values.
+    virtual void clear_image() override;
 
   protected:
-    foundation::Image*  m_filter_image;
-
-    void create_image(
+    virtual void create_image(
         const size_t    canvas_width,
         const size_t    canvas_height,
         const size_t    tile_width,
